inventory_project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
psycopg2-binary>=2.9,<3.0
django-filter>=23.0,<24.0
django-cors-headers>=4.0,<5.0
